requirements.txt
aiogram==2.11.2
aiogram-dialog==0.11.1
aiohttp==3.7.3
aioredis==1.3.1
aiosqlite==0.16.1
anyio==2.1.0
APScheduler==3.7.0
async-timeout==3.0.1
asyncpg==0.21.0
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
emoji==1.2.0
environs==8.0.0
greenlet==1.0.0
hiredis==1.1.0
idna==3.1
iso8601==0.1.14
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
marshmallow==3.10.0
matplotlib==3.4.2
multidict==5.1.0
networkx==2.5.1
numpy==1.20.3
parse==1.18.0
Pillow==8.2.0
pycryptodome==3.10.1
pyparsing==2.4.7
PyPika==0.44.1
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqtgraph==0.12.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
qiwipy==2.1.6
requests==2.25.1
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.4.0b2
tortoise-orm==0.16.21
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
yarl==1.6.3
validate-email